feat(260409-0l6): move recent entries to card level, add beverage quick-set sizes
- Move RecentEntriesList from water-tab to liquids-card so it's visible on all tabs
- Add quick-set size buttons (40, 200, 330, 500ml) to beverage tab
- Remove recent entries infrastructure from water-tab (hooks, state, dialog)
