+++ src/components/liquids/beverage-tab.tsx
@@ -119,6 +119,24 @@ export function BeverageTab() {
         className="h-10 mb-3"
       />
 
+      {/* Quick-set size buttons */}
+      <div className="flex gap-2 mb-3">
+        {[40, 200, 330, 500].map((size) => (
+          <Button
+            key={size}
+            variant="outline"
+            size="sm"
+            onClick={() => setPendingAmount(size)}
+            className={cn(
+              "flex-1",
+              pendingAmount === size && theme.activeToggle
+            )}
+          >
+            {size}
+          </Button>
+        ))}
+      </div>
+
       {/* Input Controls */}
       <div className="flex items-center justify-between gap-3">
         {/* Decrement Button */}
